+++ src/app/services/wordlist.service.ts
@@ -1,8 +1,6 @@
 import { Injectable } from '@angular/core';
-import { Observable, of } from 'rxjs';
-import { map, toArray, mergeMap } from 'rxjs/operators';
+import { Observable } from 'rxjs';
 import { HttpClient } from '@angular/common/http';
-import { Map } from 'immutable';
 
 @Injectable({
   providedIn: 'root'
